+++ src/lib.rs
@@ -17,6 +17,9 @@ pub mod clock;
 pub mod help;
 pub mod interpreter;
 pub mod log;
+pub mod native;
+#[cfg(feature = "ods")]
+pub mod ods;
 pub mod output;
 pub mod ovm; // Bytecode execution tier
 pub mod parallel;